// frontend/src/shared/api/generated/assessments/models/attemptResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Benefit Assessments
 * OpenAPI spec version: 0.1.0
 */
import type { AssessmentInfo } from './assessmentInfo';
import type { AttemptResponseSurvey } from './attemptResponseSurvey';
import type { AttemptResult } from './attemptResult';
import type { AttemptStatus } from './attemptStatus';
import type { Grade } from './grade';

export interface AttemptResponse {
  id: string;
  assessment: AssessmentInfo;
  status: AttemptStatus;
  survey: AttemptResponseSurvey;
  claimed_grade: Grade | null;
  started_at: string;
  deadline_at: string;
  finished_at: string | null;
  time_left_seconds: number;
  total_tasks: number;
  answered_tasks: number;
  max_score: number;
  result: AttemptResult | null;
}